enhanced frontend functionalities, and modified backend

+++ frontend/src/services/api.js
@@ -33,12 +33,12 @@ const API = axios.create({
 
          //prtoducts but for admin only
         updateProduct: (id, data) =>
-          API.put(`/products/${id}`, data).then(res => res.data),
+          API.put(`/products/admin/${id}`, data).then(res => res.data),
       
         deleteProduct: (id) =>
-          API.delete(`/products/${id}`).then(res => res.data),
+          API.delete(`/products/admin/${id}`).then(res => res.data),
         addProduct: (data) =>
-            API.post("/products/add", data).then(res => res.data),
+            API.post("/products/admin/add", data).then(res => res.data),
       
         // cart
         getCart: (userId) =>
@@ -52,19 +52,23 @@ const API = axios.create({
           }).then(res => res.data),
       
         removeFromCart: (cartId) =>
-          API.delete(`/cart/remove/${cartId}`).then(res => res.data),
-      
+          API.delete(`/cart/${cartId}`).then(res => res.data),
+        updateCartQuantity: (cartId, quantity) =>
+          API.put(`/cart/${cartId}`, { quantity }).then(res => res.data),
         // orders
         checkout: (userId) =>
           API.post("/orders/checkout", { user_id: userId }).then(res => res.data),
       
         getOrderHistory: async (userId) => {
-          const res = await API.get(`/orders/user/${userId}`);
-          return res.data;
+          const res = await API.get(`/orders/${userId}`);
+          return res.data; 
         },
         // admin only prevliges
-        getAllUsers: () =>
-          API.get("/users/admin").then(res => res.data),
+        getAllUsers: () =>{
+          console.log("API HIT: /users/admin");
+  return API.get("/users/admin").then(res => res.data);
+},
+
         updateUserRole: (id, role) =>
             API.put(`/users/admin/${id}/role`, { role }).then(res => res.data),
       
@@ -72,5 +76,10 @@ const API = axios.create({
           API.delete(`/users/admin/${id}`).then(res => res.data),
 
        getAllOrders: () =>
-         API.get("/orders/admin/orders").then(res => res.data),
-      };
+         API.get("/orders/admin/viewOrders").then(res => res.data),
+  
+
+      updateOrderStatus: (orderId, status) =>
+        API.put(`/orders/admin/viewOrders/${orderId}/status`, { status }) 
+    
+    };
